Fixes for casting performance warnings

+++ Gamma/Strategy.h
@@ -680,7 +680,7 @@ namespace phsInc{
 		
 		/// Specify on/off pattern as string, e.g., "/./....."
 		Rhythm& pattern(const char* bits, char offChar='.'){
-			mSize = strlen(bits);
+			mSize = uint8_t(strlen(bits));
 			mPattern = 0;
 			for(int i=0; i<mSize; ++i){
 				mPattern |= (bits[i]!=offChar) << (mSize-1-i);
